1226981 - puppet-username and file-names insured to be non-UTF-8
* Puppet module-names are <username>-<classname>.
* <username> is restricted to ASCII-alphanumeric ([0-9a-zA-Z]) *only*
* hammer-import uses sat5-org-name as username
* If org-name is entirely characters that are not ASCII-alphanumeric,
  we were building a module-name of "-<classname>"

Fix is to notice when <username> becomes empty. In this case, we create
a unique name, "orgid<ID>", where <ID> is the Sat6 ID of the matching
Org

In testing, it was noticed that if a config-file has UTF8 characters in
its path, it results in Pulp failing to accept the uploaded module. This
issue is also addressed in this fix - similarly to the <username> fix,
remove characters that aren't [0-9a-zA-Z_\.\-], and if the resulting
string is empty, create a puppet-name of "fileid<ID>"